fs_open() accepts a signed length, fixes an infinite loop
it occureed if the path_simplify returned a negative value - instead of
being interpreter as an error it just carried on, ending up in an
infinite loop.

+++ src/init/syscalls.c
@@ -16,7 +16,7 @@ int _syscall_await(char *buf, int len) {
 	return _syscall(_SYSCALL_AWAIT, (int)buf, (int)len, 0);
 }
 
-fd_t _syscall_fs_open(const char *path, size_t len) {
+fd_t _syscall_fs_open(const char *path, int len) {
 	return _syscall(_SYSCALL_FS_OPEN, (int)path, len, 0);
 }
